usrp: Help choose uhd::time_spec_t constructor
This helps compiling with libc++ runtime library
Fixes
 /ssd2/yoe-jetson/build/tmp/work/armv8a-yoe-linux/uhd/4.4.0-r0/git/host/lib/usrp/dboard_iface.cpp:22:37: error: ambiguous conversion for functional-style cast from 'rep' (aka 'long long') to 'uhd::time_spec_t'
|         set_command_time(cmd_time + uhd::time_spec_t(time.count()));
|                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
| /ssd2/yoe-jetson/build/tmp/work/armv8a-yoe-linux/uhd/4.4.0-r0/git/host/include/uhd/types/time_spec.hpp:40:5: note: candidate constructor
|     time_spec_t(double secs = 0);
|     ^
| /ssd2/yoe-jetson/build/tmp/work/armv8a-yoe-linux/uhd/4.4.0-r0/git/host/include/uhd/types/time_spec.hpp:47:5: note: candidate constructor
|     time_spec_t(int64_t full_secs, double frac_secs = 0);
|     ^

+++ host/lib/usrp/dboard_iface.cpp
@@ -19,7 +19,7 @@ void dboard_iface::sleep(const std::chrono::nanoseconds& time)
     // FIXME: Create a delay in the FPGA on the device.
     auto cmd_time = get_command_time();
     if (cmd_time.get_real_secs() != 0.0) {
-        set_command_time(cmd_time + uhd::time_spec_t(time.count()));
+        set_command_time(cmd_time + uhd::time_spec_t(static_cast<double>(time.count())));
     } else {
         // nanosleep is not really accurate in userland and it is also not very
         // cross-platform. So just sleep for the minimum amount of time in us.
